refactor user monitoring fixer permissions scripts to use only one usecase
