+++ src/domain/entities/user-monitoring/common/UserMonitoring.ts
@@ -2,14 +2,14 @@ import { NamedRef } from "domain/entities/Base";
 import { User } from "./User";
 import { UserResponse } from "./UserResponse";
 
-export interface UserMonitoringDetails extends UserMonitoringCountResponse {
+export interface UserMonitoringExtendedResult extends UserMonitoringBasicResult {
     usersBackup: User[];
     usersFixed: User[];
     eventid: string;
     userProcessed: UserResponse[];
 }
 
-export type UserMonitoringCountResponse = {
+export type UserMonitoringBasicResult = {
     invalidUsersCount: number;
     listOfAffectedUsers: NamedRef[];
     response: string;
