+++ src/domain/entities/user-monitoring/common/UserWithoutTwoFactor.ts
@@ -3,5 +3,4 @@ import { NamedRef } from "domain/entities/Base";
 export type UserWithoutTwoFactor = {
     invalidUsersCount: number;
     listOfAffectedUsers: NamedRef[];
-    response: string;
 };
