+++ src/domain/entities/user-monitoring/two-factor-monitoring/TwoFactorUserOptions.ts
@@ -1,8 +1,8 @@
 import { NamedRef } from "domain/entities/Base";
-import { UserMonitoringCountResponse } from "../common/UserMonitoring";
+import { UserMonitoringBasicResult } from "../common/UserMonitoring";
 
 export interface TwoFactorUserOptions {
-    userGroupsResponse?: UserMonitoringCountResponse;
+    userGroupsResponse?: UserMonitoringBasicResult;
     pushProgramId: NamedRef;
     twoFactorGroup: NamedRef;
 }
