--- src/domain/usecases/user-monitoring/permission-fixer/GetPermissionFixerConfigUseCase.ts
@@ -1,9 +1,0 @@
-import { PermissionFixerConfigRepository } from "domain/repositories/user-monitoring/permission-fixer/PermissionFixerConfigRepository";
-
-export class GetPermissionFixerConfigUseCase {
-    constructor(private configRepository: PermissionFixerConfigRepository) {}
-
-    async execute() {
-        return this.configRepository.get();
-    }
-}
